Completed stretch assignment, kind of.
I modified the final game to make a little more 'sense'. I removed clicking
pushes ship to the top mechanic because it interferes with mousedowns causing
movement, unless we are meant to check and only fire that movement when it is
the top ship. While I am completely capable of implementing that, it didn't seem
useful.

Additionally, instead of a mouseup event causing the ship to move back, a
mouseleave event does, so that there is a game in trying to keep your mouse
inside the ship.